requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1
django-filter==22.1
django-heroku==0.3.1
gunicorn==20.1.0
psycopg2==2.9.3
python-decouple==3.6
sqlparse==0.4.2
tzdata==2022.2
whitenoise==6.2.0
